Refactoring ancient trash code. Splitting code base into components. No new features.
