+++ css/_weapon-pop-up.css
@@ -47,11 +47,16 @@
 			text-align: left;
 		}
 
-		&:not(:first-child):hover,
+		/* &:not(:first-child):hover,
 		.table-header-entry:not(:first-child):hover {
 			background-color: rgba(255, 255, 255, 0.05);
 			cursor: pointer;
-		}
+		} */
+	}
+	.table-body .table-entries:hover,
+	.table-entries .table-header-entry:not(:first-child):hover {
+		background-color: rgba(255, 255, 255, 0.05);
+		cursor: pointer;
 	}
 }
