+++ script/components/Fighters/FighterSpritesView.js
@@ -1,39 +1,39 @@
 import classNames from "../../helper/DomClassNames.js";
 import * as URL from "../../helper/url.js";
 
-class FighterSprites {
+class FighterSpritesView {
 	// Draw the sprite | Sprites can be drawn up to 6 in the same component
 	// I then use CSS to position them correctly.
-	#fighterSpriteMarkUp = (fighterNumber, fighterMax, currentWeaponSprite) => {
+	#fighterSpriteMarkup = (fighterNumber, fighterMax, currentWeaponSprite) => {
 		const spriteLocation = `./${URL.DATA}/${currentWeaponSprite}`;
 		return `<img src="${spriteLocation}" alt="fighter sprite" class="${classNames.fighterSprite} ${classNames.fighterSprite}--${fighterNumber} ${classNames.fighterSpritesMax}--${fighterMax}" />`;
 	};
 
 	// Draw sprites equal to number of fighters
-	#fighterArrayMarkUp = (currentFighterObject) => {
-		const currentWeaponSprite =
-			currentFighterObject.additionalFighterData.spriteName;
-		const maxFighters = currentFighterObject.num;
-
+	#fighterArrayMarkup = (currentFighterObject) =>
 		// Create an Array from a number, and inject Single fighterSprites equal to fighter Num (wing size)
-		return Array.from(
-			{ length: maxFighters },
+		Array.from(
+			{ length: currentFighterObject.num },
 			(_, i) =>
-				`${this.#fighterSpriteMarkUp(i + 1, maxFighters, currentWeaponSprite)}`
+				`${this.#fighterSpriteMarkup(
+					i + 1,
+					currentFighterObject.num,
+					currentFighterObject.additionalFighterData.spriteName
+				)}`
 		).join("");
-	};
-	#fighterCostMarUp = (currentFighterObject) =>
+
+	#fighterCostMarkup = (currentFighterObject) =>
 		`<p class="${classNames.fighterSpriteCost}">${currentFighterObject.op_cost}</p>`;
 
 	render(currentFighterObject) {
 		const markup = `
 					<div class="${classNames.weaponSpriteParent}">
 						<div class="${classNames.fighterSpriteContainer}">
-                            ${this.#fighterArrayMarkUp(currentFighterObject)}
+                            ${this.#fighterArrayMarkup(currentFighterObject)}
 						</div>
-						${this.#fighterCostMarUp(currentFighterObject)}
+						${this.#fighterCostMarkup(currentFighterObject)}
 					</div>`;
 		return markup;
 	}
 }
-export default new FighterSprites();
+export default new FighterSpritesView();
