--- script/components/Weapons/Weapon.js
@@ -1,2 +1,0 @@
-class Weapon {}
-export default new Weapon();
