+++ script/helper/DataSet.js
@@ -4,5 +4,6 @@ const DataSet = {
 	dataCategory: "data-category",
 	dataArc: "data-arc",
 	dataAngle: "data-angle",
+	dataButtonValue: "data-button-value",
 };
 export default DataSet;
