+++ script/helper/altTextLib.js
@@ -1,0 +1,8 @@
+// Library to keep all alt text
+// from images for example
+
+const altTextLib = {
+	turrentGunSprite: "Turrent Gun Sprite",
+	weaponBaseSprite: "Weapon Base Sprite",
+};
+export default altTextLib;
